# Update `BezierPath` constructor